disable parallel test execution
