+++ src/test/scala/spark/dicom/deidentifier/TestDicomDeidentifier.scala
@@ -24,16 +24,25 @@ import ai.kaiko.dicom.DicomDeidentifyDictionary.{
   EMPTY_STRING,
   DUMMY_STRING
 }
-import ai.kaiko.spark.WithSpark
 import org.apache.log4j.Level
 import org.apache.log4j.LogManager
 import org.dcm4che3.data.Keyword.{valueOf => keywordOf}
 import org.dcm4che3.data._
 import org.scalatest.CancelAfterFailure
-import org.scalatest.DoNotDiscover
 import org.scalatest.funspec.AnyFunSpec
 import java.io.File
 
+import org.apache.spark.sql.SparkSession
+import org.apache.log4j.Level
+import org.scalatest.BeforeAndAfterAll
+
+trait WithSpark {
+  var spark = {
+    val spark = SparkSession.builder.master("local").getOrCreate
+    spark.sparkContext.setLogLevel(Level.ERROR.toString())
+    spark
+  }
+}
 
 object TestDicomDeidentifier {
   val SOME_DICOM_FILEPATH =
@@ -45,17 +54,21 @@ object TestDicomDeidentifier {
   }
 }
 
-@DoNotDiscover
 class TestDicomDeidentifier
     extends AnyFunSpec
-    with WithSpark
-    with CancelAfterFailure {
+    with BeforeAndAfterAll
+    with CancelAfterFailure
+    with WithSpark {
   import TestDicomDeidentifier._
 
   val logger = {
     val logger = LogManager.getLogger(getClass.getName);
     logger.setLevel(Level.DEBUG)
     logger
+  }
+
+  override protected def afterAll(): Unit = {
+    spark.stop
   }
 
   val SOME_DROPPED_COL = keywordOf(Tag.ContainerComponentID)
